refactor: create App struct
