+++ crates/tuono/src/lib.rs
@@ -1,5 +1,7 @@
+mod app;
 pub mod cli;
+mod mode;
+mod route;
 mod scaffold_project;
 mod source_builder;
-mod ssg;
 mod watch;
