--- crates/tuono/src/ssg.rs
@@ -1,8 +1,0 @@
-pub struct SSG {}
-
-impl SSG {
-    pub fn new() -> Self {
-        SSG {}
-    }
-}
-
